% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_relative_expression.r
\name{Calculate_relative_exp}
\alias{Calculate_relative_exp}
\title{Calculate the relative expression using delta delta ct method}
\usage{
Calculate_relative_exp(data_wide, internal_ctrl, format = "long")
}
\arguments{
\item{internal_ctrl}{GAPDH, ACTIN, etc}

\item{format}{"long" or "wide"}

\item{data}{a data frame with three columns: sample_name, gene_name, ct}
}
\value{
a long or wide form data frame
}
\description{
Calculate the relative expression using delta delta ct method
}
\examples{

}
