% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/delta_ct.R
\name{delta_ct_chip}
\alias{delta_ct_chip}
\title{Calculate Delta Ct For ChIP samples}
\usage{
delta_ct_chip(df, col, control = "Control")
}
\arguments{
\item{df}{data frame must have columns Well, Cq, Sample, Group, Primer.}

\item{col}{column number or character that contains the control group}

\item{control}{Group name of the control sample - defaults to control}
}
\description{
This function takes a data frame and calculates the 2^(ct_control - ct_experiment)
}

