% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/delta_ct.R
\name{delta_ct_rt}
\alias{delta_ct_rt}
\title{Calculate Delta Ct of RT-qPCR compared to a meta gene}
\usage{
delta_ct_rt(df, control = "meta")
}
\arguments{
\item{df}{data frame must have columns Well, Cq, Sample, Group, Primer}

\item{control}{name of control primer sample defaults to 'meta'}
}
\description{
This function takes a data frame and calcualte  the delta ct
compared to a normalization gene or a meta gene gene
}

