% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.r
\name{import_data}
\alias{import_data}
\title{Title  import the raw data file  and infer the replicates and the ct range.}
\usage{
import_data(file)
}
\arguments{
\item{file}{the text file that exported from sds software}
}
\value{
a list containing a wide form and a long form dataframe
}
\description{
Title  import the raw data file  and infer the replicates and the ct range.
}
\examples{

}
