% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/meta_control.R
\name{meta_control}
\alias{meta_control}
\title{Geometric mean of control genes for normalizing}
\usage{
meta_control(df, control)
}
\arguments{
\item{df}{data frame containing all data, from the process_data function, containing columns, Primer, Sample, Group, mean, sd}

\item{control}{control genes to calculate the geometric mean}
}
\description{
Takes a data frame and the names of control primers, can be more than one, and returns a data frame
that contains the geometric mean of the control primers - allowing for relative values to be calculated. Similiar to other
functions in this package, requires a specific data frame
}

