% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/negative_control.R
\name{negative_control}
\alias{negative_control}
\title{Negative Control}
\usage{
negative_control(df, primer)
}
\arguments{
\item{df}{data frame must have columns Well, Cq, Sample, Group, Primer.}

\item{Primer}{to compare values to}
}
\description{
Compare values to a negative control locus
}

