% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process data}
\usage{
process_data(df, threshold = 0.5)
}
\arguments{
\item{df}{data frame must have columns Well, Cq, Sample, Group, Primer.}

\item{threshold}{standard deviation for replicates over which a warning will be flagged
default = 0.5}

\item{control_group}{Sample name to compare data to}

\item{control_primer}{Primer to compare enrichment against}
}
\description{
Take a data frame containing qPCR data and return a data frame with normalized and relative data
}

