% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_view_ct.r
\name{qview_data}
\alias{qview_data}
\title{Title Create an overview of the Ct value of all samples}
\usage{
qview_data(data_long, ct_diff = 0.5)
}
\arguments{
\item{data_long}{long data frame from the output of import_data function}

\item{ct_diff}{the threshold for big variation of the replicate wells, the default is 0.5}
}
\value{
a ggplotly object showing in row and column with tooltip(ct, well id, different ct)
}
\description{
Title Create an overview of the Ct value of all samples
}
\examples{

}
