% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/rel_to.R
\name{rel_to}
\alias{rel_to}
\title{Relative To}
\usage{
rel_to(df, primer)
}
\arguments{
\item{df}{data frame must have columns Well, Cq, Sample, Group, Primer.}

\item{primer}{What is the name of the control primer - should be in the 'Primer' column of df}
}
\description{
Compare qPCR data to a control primer
}

