% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/rel_to.R
\name{rel_to_primer}
\alias{rel_to_primer}
\title{Relative To Control Primer}
\usage{
rel_to_primer(df, primer)
}
\arguments{
\item{df}{data frame must have columns Well, Cq, Sample, Group, Primer.}

\item{primer}{What is the name of the control primer - should be in the 'Primer' column of df}
}
\description{
Compare qPCR data to a control locus
For example: take the normalized data from delta_ct and compare to a negative locuse.
Function should be used as second step in ChIP-qPCR
}

