% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/rel_to.R
\name{rel_to_sample}
\alias{rel_to_sample}
\title{Relative to Control Sample}
\usage{
rel_to_sample(df, col, sample)
}
\arguments{
\item{df}{data frame should contain columns  Well, Cq, Sample, Group, Primer.}

\item{col}{name or integer of column used as control sample}

\item{sample}{name of the sample to be used as the control}
}
\description{
Compare qPCR data to a control sample.
For example: take normalized data and get relative fold enrichment
Function should be used as first step in qRT-PCR
}

